Enter will add a numeric locator without debounce
If one typed a numeric locator after adding a bubble,
Enter will add it to the last-added bubble without
waiting for it to happen automatically after debounce.

It may happen that one will press Enter intending to
confirm the numeric locator right as the debounce
finished. In that case, one  would press Enter
after the locator is moved and confirm the dialog
unintentionally. To prevent it, Enter in bubble-input
will not accept the dialog for 500ms after the locator
is transferred to the bubble.