selected/open sections refactoring
- selected and open items will include top-level items
that have notes. Child notes of top-level items will
then also be counted as "opened" or "selected" so they
are subject to the same search logic
- in library mode, selected items are split into
"Selected notes" and "Selected top level items".
Clicking on a top-level item will select the first
note of the top-level item in the itemTree.
- in list mode, selected items appear in one group
because top-level items are displayed either way.
All children of top-level selected items are included.