enhance copy between libraries, linked collections
- allow one to copy a linked collection of one library
into another linked collection after the first copy.
This will add new subcollections and items (including child attachments,
notes, and annotations) into the linked collection
without editing already linked items.
- provide a new option to "replicate" a collection in
another library that will not only copy new items
and subcollections but also update metadata of items,
rearrange subcollections to match those in the source
library and delete items (including children) and collections
that are not linked. It allows one to essentially sync
the state of one collection with another.
- when a linked collection exists, context menu of
collection tree "Copy to" will only show the menu of that
linked collection. It has two menuitem options - to
copy collection or to fully replicate it
- allow one to copy a library into another library,
as well as to replicate it, same way as with collections
above
- when a collection is selected, highlight the linked
collections in other libraries with faint blue, so that
there is some indication which collections are linked
- refactoring to move the logic of copying collections/items
between libraries from collectionTree into dedicated methods
in Zotero.Libraries, Zotero.Collections and Zotero.Items.

Fixes: #3045