citation dlg: add annotation items as a mock note
Allow to insert annotations from the citation dialog
as a mock note into the document. Only library
mode is available for now.

- update editor instance to be able to create a note
from annotations without saving the note into the
database, which gives us the temporary mock note that can
be inserted from the citation dialog
- when the image or ink annotations are being inserted,
they cannot be added into the note via attachmentKey
(since it would require to save entries to database).
Instead, they are embedded as a src="data:image/..." string
- display selected annotations in a deck separate from
selected items. Clicking on a selected or opened item
will select it in the itemTree.
- adding an annotation will bubbleize it
- annotation cannot be re-sorted. They are sorted
automatically by their annotation order.
- side pane to the right of itemTree will display
a preview of selected annotations or all annotations
of a selected item. Added + icon to annotation-rows
to add annotations from the side pane.