scaffold UX tweaks
- show an alert if one clicks on 'Test Regex', 'Run detect*',
'Run do*', 'Create Web Test' buttons without having a webpage
loaded in the browser tab to guide first-time users
- add an overlay appearing on top of monaco editors
in Code and Tests tabs with instructions on how to use
Scaffold
- add a "Load URL" button next to the URL field. It is
more discoverable than just relying on Enter keypress.
- display a spinner while the webpage is loading instead
of the "Load URL" button to make it clear that loading
is in progress in case it takes a while.