new fix to failing 'should refresh on item update' (#5605)
Another potential fix to the test failure. Earlier fix
from 30784dd2413e39a3b38c017188f7873c47f13b98 seems
to not have worked.

A new explanation is that the test before it does not properly
wait for the trash to refresh before trying to select the library, in
which case collectionTree select event will be suppressed
and library selection will not happen.

Fixes: #5584