citation dlg: warn about cross-library citations
- new Zotero.Integration.refreshCitedLibraries function
to record the libraries that are currently cited in the document
- when a new item is about to be cited, show a warning
if it does not belong to currently cited libraries
- cited libraries are saved in the document preferences
so that they persist across sessions
- cited libraries are recorded as group URIs and their
names. That way, a user that does not have access to a cited
group can still receive a warning with the group's name.
Also, URIs allow us to distinguish between "My Library"-s
of different users.
- cited libraries in library mode appear at the top of the
collectionTree
- in list mode, cited libraries receive priority in search results,
so that they always appear after selected/open/cited items.